+++ lib/src/component/tab_bar.dart
@@ -111,3 +111,68 @@ class BaseTabBar extends TabBar {
                     fontSize: fontSize,
                     fontWeight: fontWeight));
 }
+
+typedef CustomTabBarItemBuilder = Widget Function(int selected, int index);
+
+class CustomTabBar extends StatefulWidget {
+  const CustomTabBar(
+      {super.key,
+      required this.itemBuilder,
+      required this.builder,
+      this.initial = 0,
+      required this.controller,
+      this.onTap});
+
+  final CustomTabBarItemBuilder itemBuilder;
+  final ValueCallbackTV<Widget, List<Widget>> builder;
+  final int initial;
+  final TabController controller;
+  final ValueCallback<int>? onTap;
+
+  @override
+  State<CustomTabBar> createState() => _CustomTabBarState();
+}
+
+class _CustomTabBarState extends State<CustomTabBar> {
+  late int selected;
+
+  @override
+  void initState() {
+    super.initState();
+    selected = widget.initial;
+    widget.controller.addListener(listener);
+  }
+
+  @override
+  void didUpdateWidget(covariant CustomTabBar oldWidget) {
+    super.didUpdateWidget(oldWidget);
+    widget.controller.removeListener(listener);
+    widget.controller.addListener(listener);
+  }
+
+  void listener() {
+    var index = widget.controller.index;
+    final maxLength = widget.controller.length;
+    if (index < 0) {
+      index = 0;
+    } else if (index > maxLength) {
+      index = maxLength;
+    }
+    if (index == selected) return;
+    selected = index;
+    setState(() {});
+  }
+
+  @override
+  void dispose() {
+    super.dispose();
+    widget.controller.removeListener(listener);
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return widget.builder(widget.controller.length.generate((int index) {
+      return widget.itemBuilder(selected, index);
+    }));
+  }
+}